operations: move Rcat tests back to main test file now S3 prob is fixed
